requirements.txt
numpy>=1.24,<1.27
scipy>=1.10,<1.12
opencv-python>=4.8,<5
matplotlib>=3.8
Pillow>=10.0
rawpy
ImageIo